requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
click==8.1.6
Django==4.2.3
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Markdown==3.4.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
psycopg2==2.9.6
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
uritemplate==4.1.1
